refactor ideas
There are a few programing concepts I want to emphasize with this refactor proposal:
1- "declarative" vs "imperative" programming
2- "separation of concerns"
3- the importance of testability.

Code is, in an ideal world, self-documenting. 
The wording that you choose to use in naming variables guides the person reading the code through what each line is doing.
This is what is meant by "declarative programming". That you see the code more as sentences an less as a bunch of for-loops, etc.

Secondly, "separation of concerns" allows for easily extending your codebase in the future, as well as making the whole of the code easier to reason about. Moreover, if a bug is found in your code, it is easier to pin-point the section of the code that needs fixing when each part of the code is split into more atomic responsibilities.

Lastly, by creating code that has strong separation of concerns, it is also easier to test the code, and test for all the edge-cases that might occur across different parts of your logic.